fix: #2 수정했던 코드 누락 추가

+++ src/main/java/com/myapp/warmwave/domain/temperture/entity/Temperature.java
@@ -15,8 +15,5 @@ public class Temperature extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-//    @OneToOne
-//    private User userId;
-
-    private Float celsius;
+    private Double celsius;
 }
